Simplified main toh app for learning purposes.
