+++ app/build.gradle
@@ -1,10 +1,10 @@
 buildscript {
-  repositories {
-    jcenter()
-  }
-  dependencies {
-    classpath 'me.tatarka:gradle-retrolambda:3.2.5'
-  }
+    repositories {
+        jcenter()
+    }
+    dependencies {
+        classpath 'me.tatarka:gradle-retrolambda:3.2.5'
+    }
 }
 
 apply plugin: 'com.android.application'
@@ -12,51 +12,35 @@ apply plugin: 'com.neenbedankt.android-apt'
 apply plugin: 'me.tatarka.retrolambda'
 
 android {
-  compileSdkVersion 24
-  buildToolsVersion "24.0.0"
+    compileSdkVersion 24
+    buildToolsVersion "24.0.1"
 
-  defaultConfig {
-    applicationId "org.charmeck.trailofhistory"
-    minSdkVersion 19
-    targetSdkVersion 23
-    versionCode 1
-    versionName "1.0"
-  }
-  buildTypes {
-    release {
-      minifyEnabled false
-      proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+    defaultConfig {
+        applicationId "org.charmeck.trailofhistory"
+        minSdkVersion 19
+        targetSdkVersion 24
+        versionCode 1
+        versionName "1.0"
     }
-  }
-  compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_8
-    targetCompatibility JavaVersion.VERSION_1_8
-  }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
-  compile fileTree(dir: 'libs', include: ['*.jar'])
-  testCompile 'junit:junit:4.12'
-  compile 'com.android.support:appcompat-v7:24.0.0'
-  compile 'com.android.support:recyclerview-v7:24.0.0'
-  compile 'com.jakewharton:butterknife:8.2.0'
-  apt 'com.jakewharton:butterknife-compiler:8.2.0'
-
-  compile 'com.jakewharton.timber:timber:4.0.1'
-
-  compile 'io.reactivex:rxjava:1.1.3'
-  compile 'io.reactivex:rxandroid:1.1.0'
-  compile 'com.jakewharton.rxbinding:rxbinding:0.4.0'
-
-  compile 'com.google.dagger:dagger:2.5'
-  apt 'com.google.dagger:dagger-compiler:2.5'
-  provided 'javax.annotation:jsr250-api:1.0'
-
-  compile 'com.squareup.okhttp3:okhttp:3.2.0'
-  compile 'com.squareup.okhttp3:logging-interceptor:3.2.0'
-  compile 'com.squareup.retrofit2:retrofit:2.0.1'
-  compile 'com.squareup.retrofit2:converter-gson:2.0.1'
-  compile 'com.google.code.gson:gson:2.6.2'
-  compile 'com.squareup.retrofit2:adapter-rxjava:2.0.1'
-  debugCompile 'com.squareup.retrofit2:retrofit-mock:2.0.1'
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    testCompile 'junit:junit:4.12'
+    compile 'com.android.support:appcompat-v7:24.2.0'
+    compile 'com.android.support:recyclerview-v7:24.2.0'
+    compile 'com.jakewharton:butterknife:8.2.0'
+    apt 'com.jakewharton:butterknife-compiler:8.2.0'
+    compile 'com.google.code.gson:gson:2.6.2'
+    compile 'com.jakewharton.timber:timber:4.0.1'
 }
