--- app/src/debug/java/org/charmeck/trailofhistory/DebugTrailOfHistoryGraph.java
@@ -1,4 +1,0 @@
-package org.charmeck.trailofhistory;
-
-public interface DebugTrailOfHistoryGraph extends TrailOfHistoryGraph {
-}
