--- app/src/debug/java/org/charmeck/trailofhistory/data/IsMockMode.java
@@ -1,9 +1,0 @@
-package org.charmeck.trailofhistory.data;
-
-import java.lang.annotation.Retention;
-import javax.inject.Qualifier;
-
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Qualifier @Retention(RUNTIME) public @interface IsMockMode {
-}
