--- app/src/main/java/org/charmeck/trailofhistory/SingleIn.java
@@ -1,7 +1,0 @@
-package org.charmeck.trailofhistory;
-
-import javax.inject.Scope;
-
-@Scope public @interface SingleIn {
-  Class<?> value();
-}
