--- app/src/main/java/org/charmeck/trailofhistory/TrailOfHistoryGraph.java
@@ -1,9 +1,0 @@
-package org.charmeck.trailofhistory;
-
-import org.charmeck.trailofhistory.data.api.TrailOfHistoryService;
-
-public interface TrailOfHistoryGraph {
-  void inject(TrailOfHistoryApp aoo);
-
-  TrailOfHistoryService trailOfHistoryService();
-}
