--- app/src/main/java/org/charmeck/trailofhistory/data/api/Results.java
@@ -1,17 +1,0 @@
-package org.charmeck.trailofhistory.data.api;
-
-import retrofit2.adapter.rxjava.Result;
-import rx.functions.Func1;
-
-public final class Results {
-  private static final Func1<Result<?>, Boolean> SUCCESS =
-      result -> !result.isError() && result.response().isSuccessful();
-
-  public static Func1<Result<?>, Boolean> isSuccess() {
-    return SUCCESS;
-  }
-
-  private Results() {
-
-  }
-}
