+++ app/src/main/java/org/charmeck/trailofhistory/poi/PointOfInterestAdapter.java
@@ -5,11 +5,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import java.util.List;
+
 import org.charmeck.trailofhistory.R;
 import org.charmeck.trailofhistory.model.PointOfInterest;
+
+import java.util.List;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
 
 /**
  * Created by Trey Robinson on 3/8/16.
