--- app/src/main/java/org/charmeck/trailofhistory/ui/main/MainActivityComponent.java
@@ -1,11 +1,0 @@
-package org.charmeck.trailofhistory.ui.main;
-
-import dagger.Component;
-import org.charmeck.trailofhistory.SingleIn;
-import org.charmeck.trailofhistory.TrailOfHistoryComponent;
-
-@SingleIn(MainActivityComponent.class)
-@Component(dependencies = TrailOfHistoryComponent.class, modules = MainActivityModule.class)
-public interface MainActivityComponent {
-  void inject(MainActivity activity);
-}
