--- app/src/main/java/org/charmeck/trailofhistory/ui/main/MainActivityModule.java
@@ -1,6 +1,0 @@
-package org.charmeck.trailofhistory.ui.main;
-
-import dagger.Module;
-
-@Module public class MainActivityModule {
-}
